Arctic.AI V6.0
This program uses OpenCV along with Pyautogui pillow and openCVzone. To use this program run with additional libraries listed in the import list. User input can be obtained by pressing "x" on the keyboard. This saves all positions on the hand in relation to every other. Along with that, it will save what message the user has requested. This works with a linear recursion mode that compares current data points to existing ones with a range determined in the function. Confidence is represented by a score of X or Y divided by 400.